requirements.txt
colorama
python-dotenv 
